upgrade lodash and drop underscore.string
If we update to the latest lodash we can use the sting functions
included and drop the usage of underscore.string.
isFinite changed from the older version of lodash to no longer coerce
strings into numbers so we need to use parseInt before checking.
parseInt on an undefined returns NaN which isFinite does not identify
as a number.
Existing tests run checks for `flying-car` definitions to be named
`FlyingCar` so there doesn’t seem to be a need to add more checks.